pyproject.toml
[project]
name = "agent-fine-tuning-trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10"
dependencies = [
    "absl-py==2.2.0",
    "accelerate==1.5.2",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.14",
    "aiosignal==1.3.2",
    "alibabacloud-credentials==0.3.6",
    "alibabacloud-darabonba-stream==0.0.2",
    "alibabacloud-gateway-spi==0.0.3",
    "alibabacloud-openapi-util==0.2.2",
    "alibabacloud-tea==0.4.2",
    "alibabacloud-tea-openapi==0.3.13",
    "alibabacloud-tea-util==0.3.13",
    "alibabacloud-tea-xml==0.0.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "babel==2.17.0",
    "backports-tarfile==1.2.0",
    "beautifulsoup4==4.13.4",
    "bitsandbytes==0.45.3",
    "bleach==6.2.0",
    "build==1.2.2.post1",
    "cachecontrol==0.14.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "cleo==2.1.0",
    "click==8.1.8",
    "colorlog==6.9.0",
    "comm==0.2.2",
    "crashtest==0.4.1",
    "cryptography==44.0.2",
    "cut-cross-entropy==25.1.1",
    "datasets==3.4.0",
    "debugpy==1.8.13",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "diffusers==0.32.2",
    "dill==0.3.8",
    "distlib==0.3.9",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "dulwich==0.22.8",
    "entrypoints==0.4",
    "et-xmlfile==2.0.0",
    "exceptiongroup==1.2.2",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "findpython==0.6.3",
    "fqdn==1.5.1",
    "frozenlist==1.5.0",
    "fsspec==2024.12.0",
    "grpcio==1.71.0",
    "h11==0.14.0",
    "hf-transfer==0.1.9",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "installer==0.7.0",
    "ipykernel==6.29.5",
    "ipython==8.34.0",
    "isoduration==20.11.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.1.0",
    "jedi==0.19.2",
    "jeepney==0.9.0",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "joblib==1.4.2",
    "json5==0.12.1",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==7.4.9",
    "jupyter-core==5.7.2",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.6",
    "jupyter-server==2.16.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.6",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "keyring==25.6.0",
    "lark==1.2.2",
    "loguru==0.7.3",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.1.3",
    "more-itertools==10.6.0",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.2.0",
    "multiprocess==0.70.16",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "notebook==7.4.5",
    "notebook-shim==0.2.4",
    "numpy==2.2.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "openai==1.68.2",
    "openpyxl==3.1.5",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pbs-installer==2025.3.17",
    "peft==0.14.0",
    "pexpect==4.9.0",
    "pillow==11.1.0",
    "pkginfo==1.12.1.2",
    "platformdirs==4.3.7",
    "poetry==2.1.1",
    "poetry-core==2.1.1",
    "prometheus-client==0.22.1",
    "prompt==0.4.1",
    "prompt-toolkit==3.0.50",
    "propcache==0.3.0",
    "protobuf==3.20.3",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==19.0.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pyparsing==3.2.1",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.3.0",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "pyzmq==24.0.1",
    "rapidfuzz==3.12.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rich==13.9.4",
    "rouge==1.0.1",
    "rpds-py==0.27.0",
    "ruamel-yaml==0.18.10",
    "ruamel-yaml-clib==0.2.12",
    "safetensors==0.5.3",
    "secretstorage==3.3.3",
    "send2trash==1.8.3",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "shtab==1.7.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "sympy==1.13.1",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "terminado==0.18.1",
    "tinycss2==1.4.0",
    "tokenizers==0.21.1",
    "tomli==2.2.1",
    "tomlkit==0.13.2",
    "torch==2.5.1",
    "torch-tb-profiler==0.4.3",
    "torchvision==0.20.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.51.3",
    "triton==3.1.0",
    "trl==0.15.2",
    "trove-classifiers==2025.3.19.19",
    "typeguard==4.4.2",
    "types-python-dateutil==2.9.0.20250809",
    "typing-extensions==4.12.2",
    "tyro==0.9.17",
    "tzdata==2025.1",
    "unsloth==2025.2.15",
    "unsloth-zoo==2025.3.15",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "virtualenv==20.29.3",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "werkzeug==3.1.3",
    "xformers==0.0.29.post1",
    "xxhash==3.5.0",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
